requirements.txt
Django==3.1.2
djangorestframework==3.11.0
requests==2.25.1